transcendence/requirements.txt
Django>=4.0,<5.0
psycopg2-binary
python-dotenv
bcrypt
Pillow

